Update: 共通化
